Correcion de backend_2

+++ src/main/java/com/troyecto/marketplace/dtos/order/OrderRequest.java
@@ -22,7 +22,7 @@ public class OrderRequest {
     private Integer totalAmount;
 
     @NotNull
-    private Integer tax;
+    private Double tax;
 
     @NotBlank
     private String currency;
